refactor(frontend): extract API from premium, historic prices, and migrate stores
Slim three stores to pure state by extracting API orchestration:
- premium store: setup/delete actions to use-premium-operations,
  capabilities watcher to use-premium-watchers
- historic prices store: cache + fetch logic to use-historic-price-cache
  (createSharedComposable), currency watcher included
- account migration store: converted entirely to composable since it
  was pure orchestration with a single consumer
