+++ frontend/app/src/components/profitloss/ReportProfitLossEventAction.vue
@@ -7,9 +7,9 @@ import AmountInput from '@/components/inputs/AmountInput.vue';
 import AssetSelect from '@/components/inputs/AssetSelect.vue';
 import DateTimePicker from '@/components/inputs/DateTimePicker.vue';
 import { useAssetPricesApi } from '@/composables/api/assets/prices';
+import { useHistoricPriceCache } from '@/modules/prices/use-historic-price-cache';
 import { usePriceTaskManager } from '@/modules/prices/use-price-task-manager';
 import { useMessageStore } from '@/store/message';
-import { useHistoricCachePriceStore } from '@/store/prices/historic';
 import { getErrorMessage } from '@/utils/error-handling';
 import { toMessages } from '@/utils/validation';
 
@@ -18,7 +18,7 @@ const { event, currency } = defineProps<{
   currency: string;
 }>();
 
-const { resetHistoricalPricesData } = useHistoricCachePriceStore();
+const { resetHistoricalPricesData } = useHistoricPriceCache();
 const { getHistoricPrice } = usePriceTaskManager();
 const { addHistoricalPrice } = useAssetPricesApi();
