+++ frontend/app/src/modules/amount-display/composables/use-asset-value.ts
@@ -1,8 +1,8 @@
 import type { ComputedRef, MaybeRefOrGetter } from 'vue';
 import { type BigNumber, Zero } from '@rotki/common';
 import { normalizeTimestamp, type Timestamp } from '@/modules/amount-display/types';
+import { useHistoricPriceCache } from '@/modules/prices/use-historic-price-cache';
 import { usePriceUtils } from '@/modules/prices/use-price-utils';
-import { useHistoricCachePriceStore } from '@/store/prices/historic';
 
 export interface UseAssetValueOptions {
   /** Asset identifier to price */
@@ -41,7 +41,7 @@ export function useAssetValue(options: UseAssetValueOptions): UseAssetValueRetur
   } = options;
 
   const { getAssetPrice } = usePriceUtils();
-  const { createKey, getHistoricPrice, getIsPending } = useHistoricCachePriceStore();
+  const { createKey, getHistoricPrice, getIsPending } = useHistoricPriceCache();
 
   const timestampToUse = computed<number>(() => {
     const ts = normalizeTimestamp(toValue(timestamp));
