+++ frontend/app/src/modules/history/management/forms/HistoryEventAssetPriceForm.vue
@@ -9,10 +9,10 @@ import AssetSelect from '@/components/inputs/AssetSelect.vue';
 import TwoFieldsAmountInput from '@/components/inputs/TwoFieldsAmountInput.vue';
 import { useAssetPricesApi } from '@/composables/api/assets/prices';
 import ToggleLocationLink from '@/modules/history/management/forms/common/ToggleLocationLink.vue';
+import { useHistoricPriceCache } from '@/modules/prices/use-historic-price-cache';
 import { usePriceTaskManager } from '@/modules/prices/use-price-task-manager';
 import { TaskType } from '@/modules/tasks/task-type';
 import { useTaskStore } from '@/modules/tasks/use-task-store';
-import { useHistoricCachePriceStore } from '@/store/prices/historic';
 import { useGeneralSettingsStore } from '@/store/settings/general';
 import { ApiValidationError, type ValidationErrors } from '@/types/api/errors';
 import { bigNumberifyFromRef } from '@/utils/bignumbers';
@@ -55,7 +55,7 @@ const chain = ref<string>();
 const showPriceFields = ref<boolean>(!hidePriceFields && !noPriceFields);
 
 const { useIsTaskRunning } = useTaskStore();
-const { resetHistoricalPricesData } = useHistoricCachePriceStore();
+const { resetHistoricalPricesData } = useHistoricPriceCache();
 const { getHistoricPrice } = usePriceTaskManager();
 const { currencySymbol } = storeToRefs(useGeneralSettingsStore());
 const { addHistoricalPrice } = useAssetPricesApi();
